requirements.txt
backports.functools-lru-cache==1.4
cycler==0.10.0
matplotlib==2.1.0
numpy==1.13.3
pandas==0.21.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyVCF==0.6.8
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
six==1.11.0
subprocess32==3.2.7
flask==0.12
